backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
numpy==2.2.6
pandas==2.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
gunicorn
django-filter==24.1
openpyxl==3.1.2
dj-database-url==2.2.0
psycopg[binary]==3.2.9
requests==2.32.2
djangorestframework-simplejwt==5.3.1
celery==5.3.4
redis==5.0.1
django-celery-beat>=2.6,<3
